+++ .ahub/sam/exclude.txt
@@ -26,7 +26,7 @@
 
 # Flatbuffers generated
 /ONE/runtime/libs/circle-schema/include/circle_schema_generated.h
-/ONE/runtime/onert/frontend/tflite/src/tflite_schema_generated.h
+/ONE/runtime/onert/core/src/loader/tflite_schema_generated.h
 
 # External code: Android NN API
 /ONE/runtime/nnapi-header/include/NeuralNetworks.h
