+++ runtime/onert/api/nnfw/CMakeLists.txt
@@ -11,9 +11,8 @@ set(NNFW_API_HEADERS include/nnfw.h include/nnfw_experimental.h)
 target_link_libraries(${ONERT_DEV} PUBLIC nnfw-nnapi-header)
 target_link_libraries(${ONERT_DEV} PRIVATE onert_core)
 target_link_libraries(${ONERT_DEV} PRIVATE nnfw_lib_misc)
-target_link_libraries(${ONERT_DEV} PRIVATE jsoncpp tflite_loader circle_loader ${LIB_PTHREAD})
+target_link_libraries(${ONERT_DEV} PRIVATE jsoncpp ${LIB_PTHREAD})
 target_link_libraries(${ONERT_DEV} PRIVATE trix_loader)
-target_link_libraries(${ONERT_DEV} PRIVATE traininfo_loader)
 target_link_libraries(${ONERT_DEV} PRIVATE nnfw_common)
 target_link_libraries(${ONERT_DEV} PRIVATE nnfw_coverage)
 target_link_libraries(${ONERT_DEV} PRIVATE circle_schema)
