+++ runtime/onert/core/src/loader/circle_traininfo.fbs
