+++ runtime/onert/core/src/loader/traininfo_loader.cc
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-#include "traininfo_loader.h"
+#include "loader/traininfo_loader.h"
+
 #include "circle_traininfo_generated.h"
 #include "flatbuffers/flatbuffers.h"
